requirements.txt
requests
urllib3
beautifulsoup4
lxml
playwright
cryptography
websocket-client